requirements.txt
nameko
mysql-connector-python
eventlet
